+++ packages/core/tests/python/test_webdriver.py
@@ -12,7 +12,7 @@ from . import RESULTS_DIR
 
 
 @pytest.fixture(autouse=True, scope="module")
-def disable_caching():
+def disable_caching_enable_logging():
     with mock.patch(
         "webdriver_manager.core.driver_cache.DriverCacheManager.find_driver",
         new=mock.Mock(return_value=None),
@@ -21,7 +21,16 @@ def disable_caching():
 
 
 # Tests with different Chrome versions, which trigger different download approaches.
-@pytest.mark.parametrize("version_override", [None, "114.0.5735.198", "115.0.5790.110"])
+@pytest.mark.parametrize(
+    "version_override",
+    [
+        None,  # whatever the system currently has
+        "114.0.5735",  # solveable and requires solving
+        "114.0.5735.198",  # solveable, not requiring solving, but non-existing
+        "115.0.5790.110",  # non-solveable, not requiring solving
+        "117.0.5938",  # non-solveable despite requiring solving
+    ],
+)
 def test_chrome_download(version_override):
     get_version_target = "RPA.core.webdriver.ChromeDriver.get_browser_version_from_os"
     browser_type = (