+++ docs/source/releasenotes.rst
@@ -12,7 +12,6 @@ Latest versions
 `Upcoming release <https://github.com/robocorp/rpaframework/projects/3#column-16713994>`_
 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-- Library **RPA.Robocorp.WorkItems**: Fix dirty items logging and documentation nits.
 - Library **RPA.Assistant** (:pr:`1122`; ``rpaframework-assistant`` **3.0.2**): Show
   files selected with ``Add File Input`` to make it easier for users to know it has
   been used. Also shows a warning if the image path for ``Add Image`` is not valid or
@@ -20,6 +19,14 @@ Latest versions
 
 `Released <https://pypi.org/project/rpaframework/#history>`_
 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+
+27.5.1 - 01 Nov 2023
+--------------------
+
+- Library **RPA.Browser.Selenium** (:pr:`1123`; ``rpaframework-core`` **11.2.3**): Fix
+  "There is no such driver by url" error with Chrome/Chromium versions above **114**
+  attempting a `chromedriver` version resolution.
+- Library **RPA.Robocorp.WorkItems**: Fix dirty items logging and documentation nits.
 
 27.5.0 - 24 Oct 2023
 --------------------